requirements.txt
# Core dependencies
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Vision model
torch>=2.0.0
transformers>=4.38.0

# CLI
click>=8.1.0

# Web UI
gradio>=4.0.0

